+++ lib/main.dart
@@ -1,59 +1,67 @@
 import 'dart:io';
-import 'dart:math';
 
 import 'package:args/args.dart';
-import 'package:flutter_launcher_icons/utils.dart';
+import 'package:flutter_launcher_icons/abstract_platform.dart';
 import 'package:path/path.dart' as path;
 import 'package:yaml/yaml.dart';
 import 'package:flutter_launcher_icons/android.dart' as android_launcher_icons;
 import 'package:flutter_launcher_icons/ios.dart' as ios_launcher_icons;
+import 'package:flutter_launcher_icons/web.dart' as web_launcher_icons;
 import 'package:flutter_launcher_icons/constants.dart';
 import 'package:flutter_launcher_icons/custom_exceptions.dart';
 
-const String fileOption = 'file';
-const String helpFlag = 'help';
-const String defaultConfigFile = 'flutter_launcher_icons.yaml';
-const String flavorConfigFilePattern = r'^flutter_launcher_icons-(.*).yaml$';
+const Map<String, AbstractPlatform> platforms = {
+  'web': web_launcher_icons.WebIconGenerator(),
+  'android_adaptive': android_launcher_icons.AdaptiveAndroidIconGenerator(),
+  'android': android_launcher_icons.DefaultAndroidIconGenerator(),
+  'ios': ios_launcher_icons.IOSIconGenerator(),
+};
+
 String flavorConfigFile(String flavor) => 'flutter_launcher_icons-$flavor.yaml';
 
 List<String> getFlavors() {
-  List<String> flavors = [];
-  for (var item in Directory('.').listSync()) {
+  final List<String> flavors = [];
+
+  for (FileSystemEntity item in Directory('.').listSync()) {
     if (item is File) {
-      final name = path.basename(item.path);
-      final match = RegExp(flavorConfigFilePattern).firstMatch(name);
+      final String name = path.basename(item.path);
+      final RegExpMatch match =
+          RegExp(flavorConfigFilePattern).firstMatch(name);
+
       if (match != null) {
         flavors.add(match.group(1));
       }
     }
   }
+
   return flavors;
 }
 
 Future<void> createIconsFromArguments(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
   parser.addFlag(helpFlag, abbr: 'h', help: 'Usage help', negatable: false);
+
   // Make default null to differentiate when it is explicitly set
   parser.addOption(fileOption,
       abbr: 'f', help: 'Config file (default: $defaultConfigFile)');
   final ArgResults argResults = parser.parse(arguments);
 
   if (argResults[helpFlag]) {
-    stdout.writeln('Generates icons for iOS and Android');
+    stdout.writeln('Generates icons for iOS, web, and Android');
     stdout.writeln(parser.usage);
     exit(0);
   }
 
   // Flavors manangement
-  var flavors = getFlavors();
-  var hasFlavors = flavors.isNotEmpty;
-
-  // Load the config file
-  final Map<String, dynamic> yamlConfig =
-      loadConfigFileFromArgResults(argResults, verbose: true);
+  final List<String> flavors = getFlavors();
+  final bool hasFlavors = flavors.isNotEmpty;
 
   // Create icons
-  if ( !hasFlavors ) {
+  if (!hasFlavors || argResults[fileOption] != null) {
+    // Load the config file
+    final Map<String, dynamic> yamlConfig =
+        loadConfigFileFromArgResults(argResults, verbose: true);
+
     try {
       createIconsFromConfig(yamlConfig);
     } catch (e) {
@@ -66,7 +74,8 @@ Future<void> createIconsFromArguments(List<String> arguments) async {
     try {
       for (String flavor in flavors) {
         print('\nFlavor: $flavor');
-        final Map<String, dynamic> yamlConfig = loadConfigFile(flavorConfigFile(flavor), flavorConfigFile(flavor));
+        final Map<String, dynamic> yamlConfig =
+            loadConfigFile(flavorConfigFile(flavor), flavorConfigFile(flavor));
         await createIconsFromConfig(yamlConfig, flavor);
       }
     } catch (e) {
@@ -78,40 +87,49 @@ Future<void> createIconsFromArguments(List<String> arguments) async {
   }
 }
 
-Future<void> createIconsFromConfig(Map<String, dynamic> config, [String flavor]) async {
-  if (!isImagePathInConfig(config)) {
-    throw const InvalidConfigException(errorMissingImagePath);
-  }
+Future<void> createIconsFromConfig(Map<String, dynamic> config,
+    [String flavor]) async {
   if (!hasPlatformConfig(config)) {
     throw const InvalidConfigException(errorMissingPlatform);
   }
 
-  if (isNeedingNewAndroidIcon(config) || hasAndroidAdaptiveConfig(config)) {
-    final int minSdk = android_launcher_icons.minSdk();
-    if (minSdk < 26 &&
-        hasAndroidAdaptiveConfig(config) &&
-        !hasAndroidConfig(config)) {
-      throw const InvalidConfigException(errorMissingRegularAndroid);
+  for (final AbstractPlatform platform in platforms.values) {
+    final String complaint = platform.isConfigValid(config);
+
+    if (complaint != null) {
+      throw InvalidConfigException(complaint);
     }
   }
 
-  if (isNeedingNewAndroidIcon(config)) {
-    android_launcher_icons.createDefaultIcons(config, flavor);
-  }
-  if (hasAndroidAdaptiveConfig(config)) {
-    android_launcher_icons.createAdaptiveIcons(config, flavor);
-  }
-  if (isNeedingNewIOSIcon(config)) {
-    ios_launcher_icons.createIcons(config, flavor);
-
+  for (final AbstractPlatform platform in platforms.values) {
+    if (platform.inConfig(config) &&
+        platform.logWarnings(config, out: stderr)) {
+      platform.createIcons(config, flavor);
+    }
   }
 }
 
+/// Call [loadConfigFile] with arguments inferred from [argResults].
+///
+/// If an error occurs and the (optional) [verbose] is true, log
+/// a description to stderr.
+///
+/// Treat the current working directory as [cwd] (if given), else,
+/// use './' as the current working directory.
+///
+/// Note: [cwd] was added to allow tests that require different
+///  working directories to run at the same time, without conflicting.
 Map<String, dynamic> loadConfigFileFromArgResults(ArgResults argResults,
-    {bool verbose}) {
+    {bool verbose, String cwd}) {
   verbose ??= false;
-  final String configFile = argResults[fileOption];
+  cwd ??= './';
+
+  String configFile = argResults[fileOption];
   final String fileOptionResult = argResults[fileOption];
+
+  if (configFile != null) {
+    configFile = path.join(cwd, configFile);
+  }
 
   // if icon is given, try to load icon
   if (configFile != null && configFile != defaultConfigFile) {
@@ -129,12 +147,12 @@ Map<String, dynamic> loadConfigFileFromArgResults(ArgResults argResults,
   // If none set try flutter_launcher_icons.yaml first then pubspec.yaml
   // for compatibility
   try {
-    return loadConfigFile(defaultConfigFile, fileOptionResult);
+    return loadConfigFile(path.join(cwd, defaultConfigFile), fileOptionResult);
   } catch (e) {
     // Try pubspec.yaml for compatibility
     if (configFile == null) {
       try {
-        return loadConfigFile('pubspec.yaml', fileOptionResult);
+        return loadConfigFile(path.join(cwd, 'pubspec.yaml'), fileOptionResult);
       } catch (_) {}
     }
 
@@ -169,37 +187,24 @@ Map<String, dynamic> loadConfigFile(String path, String fileOptionResult) {
   return config;
 }
 
-bool isImagePathInConfig(Map<String, dynamic> flutterIconsConfig) {
-  return flutterIconsConfig.containsKey('image_path') ||
-      (flutterIconsConfig.containsKey('image_path_android') &&
-          flutterIconsConfig.containsKey('image_path_ios'));
+bool isConfigValid(Map<String, dynamic> flutterIconsConfig) {
+  for (final AbstractPlatform platform in platforms.values) {
+    final String complaint = platform.isConfigValid(flutterIconsConfig);
+
+    if (complaint != null) {
+      return false;
+    }
+  }
+
+  return hasPlatformConfig(flutterIconsConfig);
 }
 
 bool hasPlatformConfig(Map<String, dynamic> flutterIconsConfig) {
-  return hasAndroidConfig(flutterIconsConfig) ||
-      hasIOSConfig(flutterIconsConfig);
-}
+  for (final AbstractPlatform platform in platforms.values) {
+    if (platform.inConfig(flutterIconsConfig)) {
+      return true;
+    }
+  }
 
-bool hasAndroidConfig(Map<String, dynamic> flutterLauncherIcons) {
-  return flutterLauncherIcons.containsKey('android');
-}
-
-bool isNeedingNewAndroidIcon(Map<String, dynamic> flutterLauncherIconsConfig) {
-  return hasAndroidConfig(flutterLauncherIconsConfig) &&
-      flutterLauncherIconsConfig['android'] != false;
-}
-
-bool hasAndroidAdaptiveConfig(Map<String, dynamic> flutterLauncherIconsConfig) {
-  return isNeedingNewAndroidIcon(flutterLauncherIconsConfig) &&
-      flutterLauncherIconsConfig.containsKey('adaptive_icon_background') &&
-      flutterLauncherIconsConfig.containsKey('adaptive_icon_foreground');
-}
-
-bool hasIOSConfig(Map<String, dynamic> flutterLauncherIconsConfig) {
-  return flutterLauncherIconsConfig.containsKey('ios');
-}
-
-bool isNeedingNewIOSIcon(Map<String, dynamic> flutterLauncherIconsConfig) {
-  return hasIOSConfig(flutterLauncherIconsConfig) &&
-      flutterLauncherIconsConfig['ios'] != false;
+  return false;
 }