+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.0 (**RELEASE DATE HERE**)
+
+ - Support for web favicons and launcher icons (thanks to @personalizedrefrigerator).
+ - Bug fixes and refactoring (thanks to @personalizedrefrigerator).
+
 ## 0.8.1 (2nd Oct 2020)
 
 - Fixed flavor support on windows (@slightfoot)